requirements.txt
# ML MASTER AFILIADO - Dependências
# Instale com: pip install -r requirements.txt

playwright>=1.40.0
aiohttp>=3.9.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
rich>=13.7.0
colorama>=0.4.6
schedule>=1.2.1
requests>=2.31.0
